Creation et modification DAO, request de la classe vol

+++ src/main/java/dao/DaoVol.java
@@ -2,6 +2,6 @@ package dao;
 
 import model.Vol;
 
-public interface DaoVol extends DaoGeneric<Vol, Long>{
+public interface DaoVol extends DaoGeneric<Vol, Integer>{
 
 }
